pyproject.toml
[project]
name = "pdf-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.2",
    "groq>=0.31.0",
    "pymupdf>=1.26.3",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "torch>=2.8.0",
    "transformers>=4.55.3",
]
